Validate subject-token audience in grantToken with the other token checks.
resolvePrincipalEmail now only maps an already-trusted JWT to a UC principal.
